+++ nexus-npm-repository-testsuite/src/test/java/org/sonatype/nexus/testsuite/npm/smoke/SmokeNpmIT.java
@@ -3,10 +3,8 @@ package org.sonatype.nexus.testsuite.npm.smoke;
 import java.io.File;
 
 import org.sonatype.nexus.client.core.subsystem.content.Location;
-import org.sonatype.nexus.client.core.subsystem.repository.Repositories;
 import org.sonatype.nexus.testsuite.npm.NpmITSupport;
 
-import com.bolyuba.nexus.plugin.npm.client.NpmProxyRepository;
 import com.google.common.base.Charsets;
 import com.google.common.io.Files;
 import org.junit.Test;
@@ -16,7 +14,8 @@ import static org.hamcrest.Matchers.*;
 
 /**
  * Smoke IT for NPM plugin. This IT just starts up NX with NPM plugin, creates a NPM proxy repo, and downloads
- * one single metadata file from it, and validates that URLs are properly rewritten to point back to NX instance. If any
+ * one single metadata file from it, and validates that URLs are properly rewritten to point back to NX instance. If
+ * any
  * of these fails, this IT will fail too.
  */
 public class SmokeNpmIT
@@ -31,8 +30,7 @@ public class SmokeNpmIT
   @Test
   public void smoke() throws Exception {
     // create a NPM Proxy repository that proxies mock NPM registry
-    client().getSubsystem(Repositories.class).create(NpmProxyRepository.class, REPO_ID)
-        .asProxyOf(mockRegistryServerUrl()).withName("NPM Proxy").save();
+    createNpmProxyRepository(REPO_ID);
 
     // download package root of commonjs
     final File localDirectory = util.createTempDir();
